requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy
aiosqlite
google-genai
httpx
python-dotenv
redis
pytest
pytest-asyncio
